广告机素材管理

+++ src/views/displayMgt/common/viewTwovideo.vue
@@ -18,63 +18,32 @@
                 <el-button type="primary" @click='bindList'>绑定素材</el-button>
             </el-col>
             <el-col :span='24' style='margin-top:20px'>
-                <el-table
-                    :data="arr"
-                    style="width: 100%;">
-                    <el-table-column
-                        prop="goodimage"
-                        label="商品图片"
-                        width="100"
-                        align='center'>
-                    </el-table-column>
-                    <el-table-column
-                    prop="goodname"
-                    label="商品名称"
-                    width="100"
-                    align='center'>
-                    </el-table-column>
-                    <el-table-column
-                    prop="goodtag"
-                    label="商品标签"
-                    align='center'>
-                    </el-table-column>
-                    <el-table-column
-                    prop="goodprice"
-                    label="商品价格（元）"
-                    align='center'>
-                    </el-table-column>
-                    <el-table-column
-                    prop="bindvideo"
-                    label="绑定视频"
-                    align='center'>
-                    </el-table-column>
-                    <el-table-column
-                    prop="bindaddress"
-                    label="商品详情地址"
-                    align='center'>
-                    </el-table-column>
-                    <el-table-column
-                    prop="address"
-                    label="操作"
-                    align='center'>
-                        <template scope="scope">
-                            <el-button
-                                size="small"
-                            @click="unbindVideo(scope.$index,scope.row)">解除绑定</el-button>
-                        </template>
-                    </el-table-column>
-                </el-table>
+                <el-table :data="tableData" border style="width: 100%; margin-top: 15px;">
+            <el-table-column label="商品图片" width="120" align="center">
+                <template scope="scope">
+                    <img width='50' height='50' :src="scope.row.img">
+                </template>
+    
+            </el-table-column>
+            <el-table-column prop="goodname" label="商品名称" align="center" width="200">
+            </el-table-column>
+            <el-table-column label="商品标签" width="100" align="center" prop="goodtag">
+            </el-table-column>
+            <el-table-column prop="goodprice" label="商品价格（元）" align="center" width="100">
+            </el-table-column>
+            <el-table-column prop="bindvideo" label="绑定视频" width="100" align="center">
+            </el-table-column>
+            <el-table-column prop="bindaddress" label="商品详情地址" width="300" align="center">
+            </el-table-column>
+            <el-table-column label="操作" align="center" width="200">
+                <template scope="scope">
+                    <el-button size='small' @click="deletegoods(scope.$index,scope.row.goods_id)">解除绑定</el-button>
+                    <el-button size='small'  @click="changeDiaplay(scope.$index,scope.row.goods_id)">更换素材</el-button>
+                </template>
+            </el-table-column>
+        </el-table>
             </el-col>
-            <el-dialog
-                title="提示"
-                :visible.sync="dialogDelete"
-                size="tiny">
-                <span>确认解除绑定？</span>
-                <span slot="footer" class="dialog-footer">
-                    <el-button @click="dialogDelete = false">取 消</el-button>
-                    <el-button type="primary" @click="handleDelete()">确 定</el-button>
-                </span>
-            </el-dialog>
+           
             <el-col :span='24' align='right'>
                 <div class="block">
                     <el-pagination
@@ -97,15 +66,6 @@ export default {
           video_list:{},
           dialogFormVisible:false,
           dialogDelete:false,
-          arr:[{
-            goodimage:'',
-            goodname:'',
-            goodtag:'',
-            goodprice:'',
-            bindvideo:'',
-            bindaddress:'',
-            id:''
-          }],
           options: [{
             value: '5',
             label: '5'
@@ -123,7 +83,7 @@ export default {
         this.getVideoList()
 	},
 	methods:{
-		bindList:function (str) {
+		bindList:function () {
 			this.$router.push({path:'/bindList'});
 		},
         getVideoList(){
@@ -132,38 +92,48 @@ export default {
                uid:'1209811640320002' 
             }
             updatavideo(id).then(data=>{  
-                this.video_list = data.result.video_list
-                for(var i = 0;i<this.video_list.length;i++){
-                    this.tableData.goodimage = this.video_list[i].video.goods.pic_url
-                    this.imgs = this.video_list[i].video.goods.pic_url
-                    this.tableData.goodname = this.video_list[i].video.goods.info_name
-                    this.tableData.goodtag = this.video_list[i].video.goods.tag_goods_value
-                    this.tableData.goodprice = this.video_list[i].video.goods.price_goods_value
-                    this.tableData.bindvideo = this.video_list[i].video.goods.vedio_url_url ? '是' : '否'
-                    this.tableData.bindaddress = this.video_list[i].video.goods.jump_url_url
-                    this.tableData.id = this.video_list[i].video.goods.info_id
-                    if(this.tableData !== []){
-                         this.arr.push(this.tableData)
-                    }              
-                }
+                this.tableData = data.result.video_list.map(v=>{
+                    //console.log(v.video.goods.info_id)
+                    return {
+                        img:v.video.goods.pic_url,
+                        goodname:v.video.goods.info_name,
+                        goodtag:v.video.goods.tag_goods_value,
+                        goodprice:v.video.goods.price_goods_value,
+                        bindvideo:v.video.goods.vedio_url_url?'是':'否',
+                        bindaddress:v.video.goods.vedio_url_url,
+                        id:v.video.goods.info_id
+                    }
+                });
             })
         },
         //解除绑定
-        unbindVideo(index,row){
+        deletegoods(index,row){
             this.dialogDelete = true;
             this.inx = index
         },
-        handleDelete(){
+        deletegoods(index,goods_id){
             //确认解除
-            this.dialogDelete = false;
-            let banner_id =this.arr[this.inx].id
-            let banId = {
-                goods_id:banner_id
-            }
-            deleteBindGoods(banId).then(data => {
-                //console.log(data)
-                this.arr.splice(this.inx,1);
-            })  
+            this.$confirm('删除该商品后将无法撤回，是否继续', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() =>{
+                goods_id = this.tableData[index].id
+                let banId = {
+                   goods_id:goods_id
+                }
+                deleteBindGoods(banId).then(data => {
+                    this.tableData.splice(this.inx,1)
+                    this.getVideoList()
+                })  
+            })
+         },
+         changeDiaplay(index){
+            let id= this.tableData[index].id
+            //console.log(id)
+            const path = '/bindList?goods_id=' + id;
+            this.$router.push({ path: path });
+			//this.$router.push({path:'/bindList',hash:id});
          }
 
 	}
